requirements.txt
pyTelegramBotAPI
requests
pypdf
sentence-transformers
numpy
langgraph
langchain-core
langchain-community
langchain-openai
langchain-text-splitters
apscheduler
haystack-ai
